Refactor: Extract router configuration to App
Not sure if this new layer is meaningful, but it feels good to separate
out this kind of configuration from the index layer, and I guess in
principle we could write some integration tests at this level?
